+++ libs/cork/src/util/timer.cpp
@@ -31,7 +31,7 @@
 void gettimeofday(struct timeval* t,void* timezone)
 {       struct _timeb timebuffer;
         _ftime( &timebuffer );
-        t->tv_sec=timebuffer.time;
+        t->tv_sec=static_cast<long>(timebuffer.time);
         t->tv_usec=1000*timebuffer.millitm;
 }
 #else
